Rename selected_grammar to selected
